Added tracker vectors for inputs and outputs.
Now used inputs and outputs are tracked to be able to detect unused ones
and force some restrictions on output connections. Outputs are only
allowed to be connected to one and only one gate. Gates are only allowed
to connect to one and only one output and any quantity of other gates.