Validation of graphs.
Implemented validation of circuits. Arity is allowed to be zero but
forbidden on the validation. We validate the ariry of the gate is not
zero and there are always the same incoming connections as arity. Also
DAG is enforced on all independent sub-circuits. Reachabilty checks are
also inplace to prevent any orphan gate and any gate that is not in the
path of a input-output flow, that meaning that all gates need to consume
eventually input or input-derived values and produce ouput values or
values that will eventually compute an output.