Optimized gate entry layout and remove smallvec.
Removed smallvec dependency for now. Use normal vector at the moment.
Changed test names on the builder to account for the new name. Changed
layout on the builder to store a vec of tuples with the gates and edges
instead of 3 vectors of gates and edges. This should improve locality.