Analyzer setup for cricuit analyses.
Added a Analyzer that can be used from a future optimizer so
optimization passes can obtain information about the circuit they are
trying to modify. Allows for invalidation of information not preserved
after analyses and for recursive analyses computation on complex ones.
Mantained the topological analysis so one can obtain the vector of
indexes of gates on topological order.