Clone and Drop semantics.
Implemented new internal nodes for clone and drop of circuit values.

Clone nodes will be required when multiple moves are needed on one
value. The clone borrows read-only the value and produces enough clones
for all of them. Topologically, should be inserted before all the
consuming nodes of that value, but this should be guaranteed as clone
nodes are implemented as a special type of gate with only one input and
multiple outputs. Input and output denpendency is still handled well.

Drops are needed only when all usages of a value are borrows, so we need
special handling of the topological positioning. Drops need to be placed
always after the nodes that borrow the value dropped.

Still need to do the logic of clone/drop node placement, but analyses
should already handle this nodes well according to their semantics.