Removed Drop.
For the moment, I think drop is not necessary, as a future wire allocator
could manage the overwrite of already used / only borrowed values. This
way we also allow for the runtime executors to overwrite values not by
dropping them, but by reusing the resources and writing over the memory
buffers already allocated. This would need clarification and expansion,
but for now, it is OK.